added savings table

+++ Client/Models/ExpensesModel.cs
@@ -3,15 +3,15 @@
 public class ExpensesModel
 {
     public int Id { get; set; }
-    public double Housing { get; set; }
-    public double Groceries { get; set; }
-    public double Utilities { get; set; }
-    public double Vacation { get; set; }
-    public double Transportation { get; set; }
-    public double Medicine { get; set; }
-    public double Clothing { get; set; }
-    public double Media { get; set; }
-    public double Insuranses { get; set; }
+    public float Housing { get; set; }
+    public float Groceries { get; set; }
+    public float Utilities { get; set; }
+    public float Vacation { get; set; }
+    public float Transportation { get; set; }
+    public float Medicine { get; set; }
+    public float Clothing { get; set; }
+    public float Media { get; set; }
+    public float Insuranses { get; set; }
     public DateTime Date { get; set; }
     public int MonthId { get; set; }
 }
