+++ Client/Models/IncomeModel.cs
@@ -3,8 +3,8 @@
 public class IncomeModel
 {
     public int Id { get; set; }
-    public double Employment { get; set; }
-    public double SideHustle { get; set; } // paritial empoyment
-    public double Dividends { get; set; }
+    public float Employment { get; set; }
+    public float SideHustle { get; set; } // paritial empoyment
+    public float Dividends { get; set; }
     public DateTime Date { get; set; }
 }
