+++ Client/Models/SavingsModel.cs
@@ -3,10 +3,10 @@
 public class SavingsModel
 {
     public int Id { get; set; }
-    public double EmergencyFund { get; set; }
-    public double RetirementAccount { get; set; }
-    public double Vacation { get; set; }
-    public double HealthNeeds { get; set; }
+    public float EmergencyFund { get; set; }
+    public float RetirementAccount { get; set; }
+    public float Vacation { get; set; }
+    public float HealthNeeds { get; set; }
     public DateTime Date { get; set; }
     public int MonthId { get; set; }
 }
